new arcade alg

+++ include/main.h
@@ -73,7 +73,7 @@ void initialize(void);
 void disabled(void);
 void competition_initialize(void);
 void opcontrol(void);
-std::vector<int> arcadeControl();
+std::vector<float> arcadeControl();
 #ifdef __cplusplus
 }
 #endif
